separating application into a several files, adding MovieFactory struct for creating new movies
