+++ observer.go
@@ -1,0 +1,5 @@
+package main
+
+type Observer interface {
+	Update(user User)
+}
