+++ user.go
@@ -1,0 +1,6 @@
+package main
+
+type User struct {
+	Username string
+	Password string
+}
